Change the constructor to an explicit ‘new’ method
So I am not as clever as I hoped to be.  Trying to abuse the __call()
metamethod was leading to gettable() loops as Lua tried to find
properties when we would attempt to pass in parameters to overwrite
the defaults of things like ‘background_color’.  This is actually a
problem with how I was using __index() but is nonetheless a sign that
I should step back and simplify the code for the approach.

This patch removes the __call() metamethod and changes our code to use
LNVL.Scene:new() instead.  And we set default background color in the
process, as I was using this to test in passing arguments to the
constructor to make sure they affected anything.