Implement a method for drawing text on scenes
This patch adds the LNVL.Scene:drawText() method.  It uses two new
properties we add to the class:

1. foreground_color
2. font

Right now these are basic: black and a 20pt Sans Vera, respectively.
But it gives us a way to draw text inside of the scene container that
properly wraps.  At this size we can fit a decent amount of dialog
into one scene container.