Fix a bug with loops in LNVL.Scene
Simply setting the __index() metamethod to LNVL.Scene itself creates
loops when we go to lookup things such as the ‘contents’ property.
However, we need it in order to find our class methods.  So this patch
changes the function to perform a rawget() on the table, i.e. object,
for anything that is not a function.  If it is a function then we get
it directly from LNVL.Scene and return that.  This seems to clear up
our infinite loop problem.  But probably not since this whole mess has
demonstrated my misunderstanding of __index() and metatables in
general with Lua.