Remove the extra definition of LNVL.Character.__index()
The code for LNVL.Character accidentally has two definitions for the
__index() metamethod.  The second one uses the
LNVL.possiblyGetClassMethod() utility which commit 7532cdd8 removes.
So we need to delete that function call and replace it with a simple
rawget() instead.  That lets us access the special ‘currentDialog’
property while still falling back on LNVL.Character.

Conflicts:
	src/character.lua