Redefine the structure of LNVL instructions
As I began to write the code itself I realized the existing design has
flaws.  Some of the properties for instructions were not necessary.
And the overall design was not as practical as I thought it would be.
This patch removes the example instruction because it no longer fits
with the new definition.  The patch also simplifies the definition of
the ‘say’ instruction, which is the one I am implementing to test the
practicality of this design.