Convert LNVL.Scene contents into instructions
This patch modifies the constructor for LNVL.Scene so that it fills
its ‘contents’ array with information naming LNVL instructions, as in
objects of the LNVL.Instruction class.  Then when drawing content to
screen we just fetch the appropriate instruction based on that content
information and execute it.

The only instruction we have at the moment is ‘say’, and after this
patch the engine does nothing that it did not do before.  So thus far
we are back where we began except with a more complex flow control.
But it will allow for easier extension later on because we should no
longer have to modify drawCurrentContent(), at least not in any
significant way.

Granted, I’ll eat those words later.