Replace ‘Scene.contents’ with ‘Scene.opcodes’
This patch rips out the existing system of treating LNVL.Scene
contents as meta-information for LNVL instructions and instead
replaces those contents with opcodes, i.e. a more direct (and useful)
representation of the information we give to the instruction handling
part of the engine.  Admittedly, this is again one of those changes
which provides no new functionality, but paves the way for future
flexibility.