Return ‘say’ opcodes from LNVL.Character:says() instead of characters
This patch changes the way we send character dialog through the
engine.  When we call says() on a character we now return an opcode
containing the dialog and a reference to the LNVL.Character object,
for future reference.  We store that opcode in a list of opcodes for
the LNVL.Scene in which we’re using the character.

Because we are creating opcodes that contain each line of dialog and
storing those in a list we no longer need the dialog list that
character objects themselves have.  So this patch also deletes the
‘dialog’ and ‘dialogIndex’ properties of characters, as we do not use
them as part of this new execution flow.

However, with this patch in place the engine will fail to run the
second example script.  It causes an error by which the says() method
of the character objects is not found.