Fix ‘say’ instructions created by characters not working
When we encounter the method Character:says() in a script we create
the opcode for the ‘say’ instruction.  At the time we create the
opcode we have no access to the Scene object in which the character is
speaking; but we do later when we loop through all of the opcodes
created within that scene, because there we will find the opcode that
says() created.  In order for the actual dialog to render to screen
the ‘say’ instruction will need a Scene object, otherwise it will not
know where to send the text.  So when processing all of the opcodes
for a scene we check to see if any need a ‘scene’ property, and if so
add it accordingly.  Except we are doing this incorrectly.  We need to
put the ‘scene’ property into the ‘arguments’ of the opcode, and not
the opcode itself.  This patch corrects that bug, and with this patch
the second example script works as intended.