Begin refactoring opcode processing for scenes
This patch creates the Scene:createOpcodeFromContent() method and uses
that to replace the large for-loop that was growing in the Scene
constructor.  However, this patch breaks some existing functionality.
For one, it does not carry over the support for the ‘monologue’
opcode.  It also prevents the example scripts for running because by
the time we call Scene:drawContainer() the Scene object no longer has
a valid ‘backgroundColor’ property.