Fix ‘say’ opcodes not having access to Scene objects
This patch sets the Scene for the opcode inside of
drawCurrentContent() instead of when we are converting Scene content
into opcodes.  I will be completely honest: I do not understand why
this delay of assignment fixes the problem; assigning the scene inside
of createOpcodeFromContent() results in a Scene object with no
properties like ‘foregroundColor’, necessary to draw dialog.