Make the Scene dimensions based on values in the Settings module
This patch attempts to provide sane default values based on the screen
width and height, drawing the dialog box on the bottom half of the
screen.  But it makes it easier for programmers to modify the default
Scene dimensions if they so choose by moving the values into the
Settings module.