Include the examples in the lnvl.love file
If we do not do this then the engine will fail to be able to find the
character images.  This is because the engine will only look within
that archive file and the save directory for the program, which we do
not set.  So we must include the examples and their images in the
archive for testing purposes.