Properly set the ‘normal’ character image
We assume that if constructor has the ‘image’ property then it will
set the ‘character.image’ property, but this is not true.  That is
because the for-loop which overrides default properties only works on
existing properties, and because ‘image’ is not pre-defined the loop
will not set its value.  Therefore we have to access the filename from
the constructor arguments directly.  This is better anyways because
now we do not have a ‘character.image’ property, which we did not
intend to use anyways.