Save the filename for the ‘normal’ character image
If we create a character with an image then we only save that image in
the ‘images’ table under the key ‘normal’.  This is fine, and useful,
but we also need to store the image under its filename for future
reference as well.  That way scripts that choose to explicitly refer
to the image will work.

This patch fixes a bug in the third example script where the engine
could not load the ‘normal’ image, precisely because we do not save
its original filename.