Fix the buggy reference to ‘self’ in LNVL.Graphics.drawText()
Originally the function was a method of LNVL.Scene and so
‘self.foregroundColor’ was available.  But we do not use objects for
the graphics code, so we must pass in the foreground color.  This
patch modifies drawText() so that it now takes three arguments; the
new second argument is the default foreground color to use.  The patch
also updates LNVL.Scene:drawText() so that it passes the correct
arguments to the updated graphics function.