Call Scene:draw() when executing the ‘say’ instruction
Currently we only call Scene:drawText().  This will never draw the
background image for the scene if it exists because only Scene:draw()
does this.  Therefore we need to call draw() to take care of general
rendering before we draw the text.  However, this patch creates a bug
in that the dialog in our sixth example script no longer appears.