Set the color mode to ‘modulate’ when drawing text
When we draw scene background images we set the color mode to
‘replace’ so that it is not affected by any other colors, allowing the
image to appear clearly.  But if we leave the color mode as ‘replace’
our dialog will appear in white, regardless of any foreground color we
try to use.  In order for dialog foreground colors to have affect we
must set the color mode back to ‘modulate’, which this patch does
inside of Graphics:drawText().