Add a utility to map opcodes to instructions
Because there is not a one-to-one mapping between opcodes and
instructions we need a way to fetch the proper instruction to execute
based on a given opcode.  Their names will often not match up, so
looking for the opcode name as a key of LNVL.Instructions is not an
option.  Instead we must use a look-up table to find the proper
instruction to run.  This patch adds a utility function with such a
look-up table and changes Scene:drawCurrentContent() to use it.