Fix Scene:createOpcodeFromContent() breaking on Character:becomes()
Calling Character:becomes() returns a table of two opcodes, one for
‘set-character-image’ and ‘draw-character’.  This causes the method
createOpcodeFromContent() to fail with an assertion error because it
assumes all tables have LNVL.Opcode for their metatable.  But this is
not true of the table which Character:becomes() returns.

We can fix this by checking to see if the ‘content’ table in
createOpcodeFromContent() has the LNVL.Opcode metatable.  If it does
not then we can assume the method which created this content returned
a table of opcodes that we must process by recursively calling
createOpcodeFromContent() on each element in that table.  This patch
adds the logic to do exactly that.

After this patch we can run the third example script, which generates
‘set-character-image’ and ‘draw-character’ opcodes, without crashing
the system.