Fix set-image instruction failing
We check the metatable of the ‘target’ in the instruction to determine
exactly how to set the image.  Right now we only support the set-image
instruction for LNVL.Character objects.  If we do not understand the
metatable we call error().  However, the code has flawed
flow-control.  Even if we do have an LNVL.Character we still call
error() in the end, causing the instruction to always fail.  What we
should be doing is only calling error() if we could not match the
metatable of the target.