Add a debugging utility to pretty-print tables
This patch creates the LNVL.Debug module that provides one function: a
utility to transform tables into pretty-printed strings.  This is
useful for debugging LNVL.  The code is the work of Julio Manuel
Fernandez-Diaz, who, in 2007, posted the function to:

    http://lua-users.org/wiki/TableSerialization

I believe he released the code into public domain, but I admit that I
am not one-hundred percent certain.