Create the ‘location’ data for draw-character opcodes
The engine crashes when running any script that calls
Character:becomes() because we do not provide the required ‘location’
property of data to the draw-image instruction.  The becomes() method
generates a draw-character opcode, which is where we want to add that
data so that the instruction has it later.  This patch adds a location
based on the value of the ‘position’ property in the opcode, which
will be equal to one of the LNVL.Position constants.  Right now we
use a hard-coded value for the distance from the screen for
LNVL.Position.Left and LNVL.Position.Right when the value really ought
to be dynamic.