Set the ‘target’ property for the ‘set-character-image’ opcode
This fixes a bug that causes a crash when executing the ‘set-image’
instruction.  That instruction expects its arguments parameter to have
a ‘target’ property representing the object whose image we are
modifying.  However, when creating the ‘set-character-image’ opcode,
which later invokes that instruction, we do not provide that target
information.  For that opcode the target needs to be the associated
Character object.  This patch fixes the problem by ensuring that is
the target in the instruction when we execute it later.