Fix a bug with first nil index property in ClampedArray:new()
We use ‘__first_nil_index’ to track where the first nil element occurs
in the array.  It has an initial value of one, which is appropriate
for empty arrays.  But we do not update the index if the constructor
receives elements to store in the array.  We must increment that index
for each value we add to the array.

With this patch in place we can run the ‘06’ example script without it
crashing.  Before it would crash after changing the scene background
because the incorrect value for ‘__first_nil_index’ caused an
out-of-bounds array access, which caused the program to pass a nil
value to code that would break the program.