Draw the scene when drawing character images
Currently when we draw a character image the rest of the screen goes
black because we do not draw the scene in that same tick.  This patch
updates the ‘draw-image’ instruction so that it always draws the
screen first, and then we draw our character image on top of that.