Modify LNVL.changeToScene() to return the ‘change-scene’ opcode
We make this change so that we can call the function within the
arguments to the Scene constructor, since anything we call there must
return an opcode for future execution instead of immediate execution;
if we changed the scene immediately it would screw up the flow of the
script.  In order to do this we must also move the definition of the
function down in the file, after all of the require() calls so that we
have access to the Opcode class.