Fix bug when finding scenes for the ‘set-scene’ instruction
When calling changeToScene() we look for the given scene name in the
global environment.  This fails if the scene we want to switch to is
not yet defined, meaning it is not accessible globally.  To fix this
we need to delay the lookup until we actually run the ‘set-scene’
instruction.  This patch changes changeToScene() to return a
‘change-scene’ opcode that contains only the name of the scene to use,
not the Scene object itself; that would still be too early to lookup
the desired scene because we create opcodes before we finish loading
all scenes from a script.  That is why we have to wait all the way
until we run ‘set-scene’ to perform the lookup, because by then every
scene will actually be available.