Fix a bug in ‘set-scene’ caused by conflict argument properties
When convert opcodes into instructions inside of the method
Scene:drawCurrentContent().  Before calling the instruction we always
add a ‘scene’ property to the arguments we give to the instruction
because it is useful to them.  However, the current definition of the
‘change-scene’ opcode uses its own ‘scene’ property.  LNVL overwrites
this property with an LNVL.Scene object, which causes the ‘set-scene’
instruction to fail.  That is because ‘set-scene’ tries to lookup the
scene in the global environment, assuming that ‘arguments.scene’ is a
string.  But this is not the case.

We fix this by renaming the ‘scene’ property in the ‘change-scene’
opcode to ‘name’.  Then the ‘set-scene’ instruction is able to look
for that scene by name.  With this patch in place we can run the ‘08’
example script with expected results.