Always draw the scene when calling drawCurrentContent()
Not every instruction renders something to screen.  When we execute
such instructions, like ‘set-image’, it results in a blank screen
because we depend on the instructions to call Scene:draw().  This
patch ensures that we call Scene:draw() before executing every
instruction so that we no longer have blank screens.  As an example,
executing the ‘03’ test script without this patch shows a black screen
every time we execute Character:becomes() because the resulting
instruction does not update the screen; that is what this fixes.