Create processor functions for existing opcodes
This patch moves all of the opcode modification code out of the
Scene:createOpcodeFromContent() method and refactors it into specific
entries in the LNVL.Opcode.Processor table.  This is not only better
organization—it keeps more of the opcode logic in the ‘opcode.lua’
file—but it also makes it possible to implement new opcodes without
having to modify any code inside of Scene:createOpcodeFromContent().