Correct the paths to documentation in two source file headers
Two files direct programmers to HTML documentation.  However, the HTML
format of the documentation is not part of the repository.  We
generate those HTML files using Pandoc and Tup.  If the programmer
does not have those two programs then those paths will lead him
nowhere.  Instead we should reference the documentation format which
is actually part of the repository, and which is readable as-is.