Fix a crash when building a scene’s opcode table
The method Scene:createOpcodeFromContent() may return nil.  That
causes a problem because we call it directly as an argument to
table.insert() so when it returns nil we end up with the problem of
calling table.insert() with not enough arguments.  We need to separate
the call to createOpcodeFromContent() and test its value so that we do
not waste space insert nil values into the opcodes table.