Test for a nil ‘arguments’ property in LNVL.Opcode.__tostring()
Not every opcode will have an arguments table.  Currently ‘no-op’ is
the only opcode that falls into this category, but others may follow.
When converting opcodes to strings we assume that table always exists
but this causes a crash when trying to convert ‘no-op’ to a string.
This patch corrects the logic of string conversion for opcodes by
adding a test for the existence of that table, and it updates the
documentation for opcodes to mention that we cannot rely on them to
always have an arguments table.