Add LNVL.Settings.DebugModeEnabled
We will use this boolean setting throughout the code to decide whether
or not to generate debugging output and data.  The default value is
true because that is what we will want up until the point we’re ready
to release a finished visual novel.

This patch also changes the order we load the LNVL modules so that
LNVL.Settings comes first.  We need to do this so that LNVL.Debug will
have access to the setting.