Test for opcode and instruction implementations in debug mode
LNVL keeps a list of opcode and instruction names that it recognizes
as valid.  If we add a new entry to that list but fail to actually
write the necessary code to implement the opcode/instruction then it
can cause some real pain-in-the-ass bugs.  This patch adds teaches
LNVL to ensure that every opcode and instruction has an implementation
on start-up.  It cannot guarantee that those implementations do what
they claim to do, but it at least protects us from forgetting to write
an implementation at all.