Prevent blank screens caused by the ‘no-op’ opcode
The method Scene:drawCurrentContent() exits early if we are processing
a ‘no-op’ to save the time of loading any instruction function.  But
this causes the screen to turn blank because we exit the method before
calling Scene:draw().  We can fix this by ensuring that we draw the
scene earlier in the method, before we possibly exit over a ‘no-op’.