Do not set a character position in LNVL.Character:becomes()
Now that the ‘draw-character’ opcode reads position information from
the Character object itself we do not need to pass that data to the
opcode directly.  This fixes a problem with positioning because our
hard-coded choice of LNVL.Position.Center in the becomes() method
would override the position property for the character anyways,
defeating the value of other methods like Character:isAt().