Generate a ‘draw-character’ opcode from LNVL.Character:isAt()
When we move a character in a script using isAt() we almost always
want to draw that character in the new spot right away.  By generating
a ‘draw-character’ opcode from this method we save script authors the
extra effort of having to explicitly tell LNVL to draw the character
after telling the engine the character’s new position.