Delay the assignment a character’s position in LNVL.Character:isAt()
We call isAt() in the constructor for scenes to move character images
around on screen.  However this fails to work properly because Lua
evaluates every function call as part of the arguments to the
constructor, which means the side-effects of every isAt() take place
before we ever begin to draw anything from the scene.  Since isAt()
changes the ‘position’ property of a character all of those changes
happen before we render the scene.  The result is that only the final
call to isAt() effects the scene.

To fix this we have to delay the assignment of the ‘position’
property.  Luckily we already implement support for this: the
‘draw-character’ opcode optionally accepts ‘position’ as part of its
arguments table.  So we need to move the assignment of the character’s
position into the opcode constructor.