Assign a Character a new position when processing ‘draw-character’
This fixes a bug from commit 65d94868 (“Delay the assignment a
character’s position in LNVL.Character:isAt()”).  We removed the
assignment of the character’s position from isAt() itself because that
was too early, but we forgot to actually assign it later at any
point.  When we receive a new position in the ‘draw-character’ opcode
we nee to assign that to the character.  If we do not then the new
position does not stay in effect and the next time we render the
character it will revert to whatever the position was before the most
recent ‘draw-character’ opcode.