Pretty-print the ‘location’ property of opcodes during debugging
Some opcodes use a ‘location’ property in their arguments as an array
of XY-coordinates.  But when we convert an opcode to a string and
print it that appears as a table memory address, which is not
helpful.  This patch will make LNVL print the coordinates as numbers
to make debugging easier.