Rename Scene:draw() to Scene:drawEssentialElements()
The draw() method should be more generic and be the public-facing
method that programs use to render scenes instead of the method
drawCurrentContent().  The latter arguably exposes too much internal
information about the scene.  Programs uses LNVL should only have to
tell a scene to draw, not *what* to draw.