Stop flattening arrays of opcodes when constructing Scene objects
Currently when the engine sees an array of opcodes it flattens those
out into the list of opcodes for a given scene so that we end up with
a continuous array of opcodes to execute one-by-one.  This approach
causes problems because there are times when we need to execute
multiple opcodes at once, e.g. drawing a character’s avatar to screen
and printing dialog from that character simultaneously.  The first
step to supporting this is to stop flattening lists of opcodes.
Instead, if we receive a list of opcodes when building a scene we will
save that list as-is.  Later when executing the contents of that scene
we will treat lists of opcodes as groups to execute all together as if
they are one, single unit.

This will only be a first step though, and this change will break the
engine until we implement some additional measures.