Draw characters on screen when they are speaking
This patch teaches the engine to (potentially) generate two opcodes
given a ‘say’ opcode:

1. draw-character
2. say

If a Character object is part of the arguments to the initial ‘say’
opcode then the result will be the two opcodes above.  This will
result in the engine later executing a ‘draw-image’ and ‘say’
instruction simultaneously, rendering the avatar of a character on
screen as that character is speaking.