Allow for optionally flattening lists of opcodes
When a scene is converting its content into opcodes it will receive
one of two things for each chunk of content:

1. A single Opcode object.
2. A table of Opcode objects.

The default behavior in the second case is to insert that table
directly into the scene’s list of opcodes.  We do this under the
assumption that the engine should treat these opcodes as a group, a
single unit to later convert into instructions and execute.  This
behavior is a good default because it takes care of such situations as
drawing a character avatar on screen at the same time he/she is
speaking.

However, there are uncommon situations where we want the engine to
flatten a table of opcodes.  The largest current example is the
‘monologue’ opcode.  This opcode results in the creation of a number
of ‘draw-character’ and ‘say’ instructions, each pair grouped together
in a table.  We screw up the engine by using that table of tables
because LNVL treats the entire monologue as a group, executing all of
those sub-opcodes at once.  That means it prints every line of the
monolgue to the screen at the same time.

Whoops.

In the case of monolgues we actually want to flatten the resulting
table of opcodes out into individual opcodes (or tables) to insert
into the scene’s opcode table.  To make that possible we will
introduce an optional ‘__flatten’ table property which opcode
processors may set, indicating to the engine that it should flatten
out that table of opcodes appropriately.