Do not draw a character in ‘say’ opcodes unless they have an image
If a ‘say’ opcode has an associated Character object then the engine
creates an additional ‘draw-character’ opcode so that the character
appears on screen along with the dialog.  This crashes the engine if
the Character object has no image to draw, i.e. the value of

    Character.images[Character.currentImage]

is nil.  So the processor for the ‘say’ opcode must check for this
before deciding whether or not to create a ‘draw-character’ opcode.