Fix the bug of not properly processing lists of opcodes in scene’s
Commit 400ec11c (‘Allow for optionally flattening lists of opcodes’)
incorrectly removes the loop from the Scene:createOpcodeFromContent()
method which deals with lists of opcodes.  We need to reintroduce that
logic to fix the engine.  But this gives us a chance to tighten it up
a little because we do not need to recursively call the method.  All
we need to do in the loop is make sure each opcode calls its process()
method, then collect and return the results.