Show all scene opcodes when in debugging mode
Currently we only show the opcodes for Scene objects which we assign
to the LNVL.currentScene global.  But that does not work.
Furthermore, we may later want to have a metatable on ‘LNVL’ for other
reasons and using one only for debugging would get in the way.  It
will be more convenient if we show the opcodes for scenes inside of
the Scene:new() constructor when using debugging mode.  That way we
are sure to see the opcodes for every scene, and it keeps
scene-related debugging code closer to the rest of the scene code.