Define the Character.borderColor property
We will use this property to control the color of the border we will
draw around character images, if the property has a non-nil value.
This will only teach the Character class to recognize the property in
its constructor like it does ‘textColor’.  The engine will not yet use
the ‘borderColor’ in any meaningful way.