Improve the vertical position of character images
Currently the placement of character images overlaps with the
container for scene dialog.  The ‘draw-character’ opcode does not take
into account the height of the image, so it is easy to create images
that spill over into the container.  We can fix that by using the
image height as part of the calculation for the vertical position.