Make character border colors transparent by default
We do not want character images to have borders by default.  However,
using a default value of nil makes it impossible to change a
character’s border color.  The constructor only lets us overwrite
properties which have non-nil values; if the ‘borderColor’ property
has a nil value by default then we can never change that in scripts.
Instead we should set the default color to be transparent.  This will
still prevent the engine from drawing a border unless we give it an
explicit color, and it will also fix the bug of being unable to change
the border color by giving it a non-nil default.