Improve the debugging output for borders in opcodes
Currently when we print out ‘draw-character’ opcodes we show a table
memory address for the optional ‘border’ argument.  It will be more
helpful to show the actual border color and width.  We can take
advantage of the tostring() support for LNVL.Color objects by
explicitly setting the metatable for the border color as well.