Improve the positioning of character images
Currently we use hard-coded values for the horizontal positioning of
character images.  We can improve this by taking into account two
factors:

1. The width of the character image itself.
2. The width and position of the scene dialog box.

We will use these two numbers to teach the ‘draw-character’ opcode to
position images relative to the left and right edges of the dialog
box, as that will look better.  And it will center images relative to
that same box.  This behavior will make LNVL more consistent with
existing visual novel engines and is (probably) a better default
choice all around.