Allow setting a Scene’s background image within the constructor
Currently the only way to set the background for a scene is to call
the setBackground() method after creating the scene.  This can appear
counter-intuitive to some people because in similar situations, such
as when creating characters, that information comes first as part of
the new() method call.

If we change the ‘background’ property for Scene objects into a string
then it becomes possible to set that property within the constructor,
allowing for scripts such as:

    START = LNVL.Scene:new{
        background = "Nice-Pretty-Background.jpg",
        "Then the dialog begins",
    }

We already have a way to load scene images given a filepath: we can
simply call setBackground() within the constructor if the script
author provides the ‘background’ property.