Remove an erroneous call to table.remove() from the Scene constructor
The second argument we give to table.remove() is incorrect.  It should
be a numeric index, not a string like we use.  But the reality is that
we do not need to call table.remove() at all.  Leaving any property
definitions inside of the ‘properties’ list will not mess up opcode
generation later when treating ‘properties’ as an array of content.
That is because when creating content we go through only the numeric
indices of ‘properties’, and that will ignore any named property
assignments automatically.