Add Scene.changeBackgroundTo() and the ‘set-scene-image’ opcode
This patch will implement the ‘set-scene-image’ opcode and provide a
function we can use in scenes to dynamically change the background
image: changeBackgroundTo().  The function is inefficient because it
will create new Image objects every time we process it, but this
provides only the essential code for an implementation.  If we need to
cache those images for optimization we will revisit this.