Fix bug where ‘set-image’ crashes when changing scene backgrounds
The instruction ‘set-image’ does not have access to an LNVL.Scene
object to modify because the ‘set-scene-image’ opcode, which generates
the instruction, does not have access to that object either.  The
problem is that we create and process the opcode before construction
of the Scene is complete.  It is not possible to give the Scene object
to the opcode processor at that time.  And we cannot delay the call to
the opcode processor either.

To get around this conundrum we modify the processor for
‘set-scene-image’ so that it passes information to ‘set-image’ to
inform it that we are modifying an LNVL.Scene object.  By the time we
actually execute the ‘set-image’ instruction we do have access to the
correct Scene object; it will be the ‘arguments.scene’ property within
that instruction code.  That instruction always works by modifying the
‘target’ property, changing one if its images to the new image.  The
opcodes which create ‘set-image’ instructions are responsible for
preparing that target.  Since ‘set-scene-image’ cannot assign the
target, i.e. the then-uncreated Scene object, we delay assignment of
the target until we execute the ‘set-image’ instruction itself.

This patch will also update the instruction documentation with a note
about where to look in the source code for details on this delayed
assignment and why we do it.