Prevent the scene opcode index from going out-of-bounds
The methods Scene:moveForward() and Scene:moveBack() do not place any
limits on incrementing and decrementing the opcode index.  Since we
call those based on user input it is possible for the user to move far
out of bounds of the actual indicies for the opcode list.  This will
not crash the game because access to the opcode list is always forced
to be within bounds, but that does not prevent the index we keep track
of from going outside of those bounds.  This patch will fix that
problem so that users cannot go too far forward or back in scenes.