Rename Scene.backgroundColor to Scene.boxBackgroundColor
This is a more accurate name.  The problem with ‘backgroundColor’ is
that the property affects something different from ‘background’.  The
former changes the color of the dialog box, while the latter changes
everything outside of that box.  So it needs to be more clear in the
code, examples, and documentation that ‘backgroundColor’ only affects
the dialog box, hence the name change.