Fix a bug where the Color module cannot open ‘src/rgb.txt’
We use the standard Lua io.open() function to read the list of colors
from disk.  This fails when we package the engine up as a demo inside
of a ‘*.love’ file because LÖVE places restrictions on how games can
access files.  It stops us from io.open() calls, even if they are in a
place that seems safe and innocent.  To fix this we need to replace
the io.open() call with the appropriate function from LÖVE’s
filesystem API.  That API adheres to the restrictions that LÖVE puts
on paths and will let the Color module correctly find files.