Configure the default scene text color from the Settings module
This will allow developers and script authors to change the global
default text color for all Scene objects.  Individual scenes will be
able to override this by defining ‘textColor’ in their constructor,
and the text colors associated with Character objects will always
override both of those when a character speaks dialog in a scene.