Define ‘image’ and ‘location’ properties for Drawable objects
This will add two properties for Drawables that we will always need:
an actual image to draw, and the coordinates indicating where on
screen to draw.  Unlike with other classes we allow the parameters to
the constructor to override properties that have default nil values,
such as the Drawable.image property.  That will let us pass in
pre-created Image objects to the constructor, which should be useful
when we start to use Drawables to represent character avatars, for
example.  We also do not intend for script authors to create Drawable
objects outside of the engine code, so we do not need to apply the
same level of ‘safety’ when it comes to modifying properties via the
constructor that we do with other classes.