Store character images as Drawable objects
This will use our Drawable class for character images instead of
LÖVE’s Image class.  It is easy to change the two places in the
Character code where we assign images to construct Drawable objects.
But we will also need to make both the ‘draw-character’ opcode and the
‘draw-image’ instruction aware of Drawables.  We can change the opcode
so that it works only with Drawables.  However, we will not be able to
do the same with the ‘draw-image’ instruction because it will continue
to receive Image objects, such as when we draw scene backgrounds.