Implement Drawable:setPosition() and use it when drawing characters
Currently the ‘draw-character’ opcode is responsible for interpreting
the meaning of LNVL.Position values in terms of how they convert to
actual, absolute screen coordinates.  We will move this logic into the
Drawable class so that moving forward we can make it possible to
position all Drawables relative to each other, not just characters.
This will be useful in the future when implementing such features as
adding arbitrary images to the foreground of scenes.

Since we are going to move that positioning logic into Drawable we
also have a chance to simplify the code and commentary for the
‘draw-character’ opcode processor.