Fix bug where ‘draw-image’ instruction crashed on border data
Now that Drawables take care of rendering borders we need to
restructure the logic in the ‘draw-image’ instruction so that it sets
the relevant properties on the Drawable object for its border.  Then
we can remove the code that draws the border from ‘draw-image’ and let
the Drawable handle all of that.