Fix a crash when scenes have transparent dialog boxes
If we set the ‘boxBackgroundColor’ of a scene to the
LNVL.Color.Transparent value then the engine will crash.  That value,
unlike the rest of the colors, is not a valid argument to the
primitive setColor() function we use before drawing the box.  We need
to test for that value before calling setColor() and then just stop
drawing if we’re asked to draw a transparent box.  We actually do not
do anything with alpha channel values; to get full transparency we
simply do not draw the dialog box at all.