Document the ‘position’ argument for the ‘draw-image’ instruction
We will implement this argument as a way to get around the issue with
Drawables not appearing in the correct position.  This will also make
it possible to pass LNVL.Position constants directly to the
‘draw-image’ instruction in a meaningful way, giving us the potential
to make it smarter at determining where to actually draw things.

Because ‘draw-image’ now accepts two arguments related to coordinates
on screen we modify the documentation to describe the relationship
between ‘location’ and ‘position’, and how ‘location’ takes precedence
in all situations.