Set the position for Drawables inside the ‘draw-image’ instruction
This fixes the bug in example script ‘11’ where the Drawable only
appears in the final position.  Originally we set the position for the
Drawable inside of the ‘draw-character’ opcode.  However, the
setPosition() method destructively modifies the coordinates where we
will eventually render the Drawable on screen.  Because we did this in
the opcode we always rendered the Drawable in the last position it has
in a scene, since we process all opcodes before rendering any content;
the position changed many times before we ever showed the image.

We fix this by taking advantage of the new ‘position’ argument to the
‘draw-image’ instruction.  By passing along that information from the
opcode to the instruction we can delay the call to setPosition() until
we are ready to actually render the Drawable.